main container enable

+++ cmd/ghostd/manager/containers.go
@@ -4,6 +4,9 @@ import (
 	"log"
 	"os/exec"
 	"sync"
+
+	"github.com/GhostNet-Dev/GhostNet-Core/internal/gconfig"
+	"github.com/GhostNet-Dev/GhostNet-Core/internal/maincontainer"
 )
 
 type Containers struct {
@@ -21,9 +24,11 @@ func (containers *Containers) GetContainer(id uint32) *Container {
 }
 
 func (containers *Containers) ReleaseContainer(id uint32) bool {
-	containers.Count--
 	if container, exist := containers.List[id]; exist {
-		container.Cmd.Process.Kill()
+		if container.Cmd != nil {
+			container.Cmd.Process.Kill()
+		}
+		containers.Count--
 		delete(containers.List, id)
 	} else {
 		return false
@@ -40,25 +45,57 @@ func (containers *Containers) CheckExistAddr(host, port string) bool {
 	return false
 }
 
-func (containers *Containers) CreateContainer(password []byte, username, host, port string) *Container {
-	wg := sync.WaitGroup{}
-	id := containers.Id
+func (containers *Containers) LoginContainer(password []byte, username, host, port string) *Container {
 	if containers.CheckExistAddr(host, port) {
 		log.Printf("already allocated port number = %s", port)
 		return nil
 	}
 
+	id := containers.Id
+	container := NewContainer(id, host, port)
+	if container == nil {
+		return nil
+	}
+	container.Username = username
+
+	container.Client.ConnectServer()
+	defer container.Client.CloseServer()
+	if !container.Client.LoginContainer(password, username, host, port) {
+		log.Println("Login Fail ", username)
+		return nil
+	}
+
+	containers.List[id] = container
+	containers.Id++
+	containers.Count++
+	return container
+}
+
+func (containers *Containers) ForkContainer(password []byte, username, host, port string) *Container {
+	if containers.CheckExistAddr(host, port) {
+		log.Printf("already allocated port number = %s", port)
+		return nil
+	}
+
+	wg := sync.WaitGroup{}
+	id := containers.Id
 	containers.Id++
 	containers.Count++
 
 	container := NewContainer(id, host, port)
+	if container == nil {
+		return nil
+	}
 	container.Username = username
 	containers.List[id] = container
 	wg.Add(1)
 
 	go func(cid uint32) {
 		log.Printf("execute node [%s] addr = %s:%s\n", username, host, port)
-		args := []string{"-p=" + port}
+		args := []string{
+			"--port=" + port,
+			"-u=" + username,
+		}
 		if host != "" {
 			args = append(args, host)
 		}
@@ -88,6 +125,52 @@ func (containers *Containers) CreateContainer(password []byte, username, host, p
 	}(id)
 
 	wg.Wait()
+	container.Client.ConnectServer()
+	defer container.Client.CloseServer()
+	if !container.Client.LoginContainer(password, username, host, port) {
+		log.Println("Login Fail ", username)
+	}
+
+	return container
+}
+
+func (containers *Containers) CreateContainer(password []byte, username, host, port string) *Container {
+	if containers.CheckExistAddr(host, port) {
+		log.Printf("already allocated port number = %s", port)
+		return nil
+	}
+
+	wg := sync.WaitGroup{}
+	id := containers.Id
+	containers.Id++
+	containers.Count++
+
+	container := NewContainer(id, host, port)
+	if container == nil {
+		return nil
+	}
+	container.Username = username
+	containers.List[id] = container
+	wg.Add(1)
+
+	go func(cid uint32) {
+		log.Printf("execute node [%s] addr = %s:%s\n", username, host, port)
+		cfg := gconfig.NewDefaultConfig()
+		cfg.Username = username
+		cfg.Password = password
+		cfg.Ip = host
+		cfg.Port = port
+		main := maincontainer.NewMainContainer(cfg)
+		main.StartContainer()
+		wg.Done()
+	}(id)
+
+	wg.Wait()
+	container.Client.ConnectServer()
+	defer container.Client.CloseServer()
+	if !container.Client.LoginContainer(password, username, host, port) {
+		log.Println("Login Fail ", username)
+	}
 
 	return container
 }
